Move RLM regression tests into test_rlm_env
